Begin JSON reporter change: Allow users to specify arbitrary JSON input/output.
This patch begins a set of changes to convert the library to use JSON as
the primary data type used to communicate with users; but for custom user
input and output.

The patch introduces the json.h header taken, and modified, from
https://github.com/nlohmann/json . The header was modified to no
longer use the namespace nlohmann, nor any macros containing NLOHMANN
so as to not interfere with other users who already incidentally
use this header.

Additionally, this patch adds `Benchmark::WithInput` to allow
the passing of arbitrary JSON as an input argument to a benchmark.
The results of which can be accessed using `State::GetInput()`.

This patch also adds `State::operator[](std::string)`, which allows
users to report arbitrary JSON as output from a benchmark. See
the documentation for examples.